proxy config: separate command from implementation
In order to make it easier to navigate and align with the other
commands, separate the code creating the cobra command from the actual
implementation in the mgrctl proxy command.
