+++ mgrctl/cmd/proxy/config_do_test.go
